backend-ai/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Google Gemini AI
google-genai==0.2.2

# Database
supabase==2.17.0

# Validation & Serialization
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0

# HTTP Client
httpx==0.28.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Utilities
loguru==0.7.3
python-dateutil==2.9.0
pytz==2024.2

# Rate Limiting
slowapi==0.1.9

# Async Support
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Development Tools
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.13.0

# Monitoring
prometheus-client==0.22.0